# api-gateway/api/docs/swagger.yaml
definitions:
...
  models.AddressReq:
    properties:
      country:
        type: string
      street:
        type: string
    type: object
  models.CreatePost:
    properties:
      description:
        type: string
      medias:
        items:
          $ref: '#/definitions/models.Media'
        type: array
      name:
        type: string
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
    type: object
  models.CustomerRegister:
    properties:
      addresses:
        items:
          $ref: '#/definitions/models.AddressReq'
        type: array
      bio:
        type: string
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
    type: object
...
  models.Error:
    properties:
      code:
        type: integer
      description:
        type: string
      error: {}
    type: object
  models.Media:
    properties:
      link:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  models.MediaUpdate:
    properties:
      id:
        type: string
      link:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  models.PostUpdate:
    properties:
      description:
        type: string
      id:
        type: string
      medias:
        items:
          $ref: '#/definitions/models.MediaUpdate'
        type: array
      name:
        type: string
    type: object
  models.Review:
    properties:
      description:
        type: string
      name:
        type: string
      rating:
        type: integer
    type: object
...
  post.Empty:
    type: object
  post.ListPostResponse:
    properties:
      posts:
        items:
          $ref: '#/definitions/post.PostSearchResp'
        type: array
    type: object
  post.Media:
    properties:
      id:
        type: string
      link:
        type: string
      name:
        type: string
      post_id:
        type: string
      type:
        type: string
    type: object
  post.PostInfo:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      medias:
        items:
          $ref: '#/definitions/post.Media'
        type: array
      name:
        type: string
      owner_id:
        type: string
      reviews:
        items:
          $ref: '#/definitions/post.Review'
        type: array
      updated_at:
        type: string
    type: object
  post.PostResp:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      medias:
        items:
          $ref: '#/definitions/post.Media'
        type: array
      name:
        type: string
      owner_id:
        type: string
      reviews:
        items:
          $ref: '#/definitions/post.Review'
        type: array
      updated_at:
        type: string
    type: object
  post.PostSearchResp:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  post.Review:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      post_id:
        type: string
      rating:
        type: integer
      udpated_at:
        type: string
    type: object
...
host: localhost:9090
info:
  contact:
    email: gofurovmurtazoxon@gmail.com
    name: Murtazoxon
    url: https://t.me/murtazokhon_gofurov
  description: This is exam server api server
  termsOfService: 2 term exam
  title: exam api
  version: "2.0"
paths:
...
  /v1/customer/register:
    post:
      consumes:
      - application/json
      description: this registers customer
      parameters:
      - description: Register customer
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CustomerRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: register customer
      tags:
      - Customer
...
  /v1/post:
    post:
      consumes:
      - application/json
      description: "post \tservice create"
      parameters:
      - description: Post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.CreatePost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.PostResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: "Create \tPost"
      tags:
      - Post
  /v1/post/delete/{id}:
    delete:
      consumes:
      - application/json
      description: "this \tdeleting information of post"
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.Empty'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: "Delete \tPost"
      tags:
      - Post
  /v1/post/get/{id}:
    get:
      consumes:
      - application/json
      description: "this \twill display the post review information"
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.PostInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: "Get \tPost Review"
      tags:
      - Post
  /v1/post/list/{page}/{limit}/{search}:
    get:
      consumes:
      - application/json
      description: "lists \tposts"
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.ListPostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: "list \tpost"
      tags:
      - Post
  /v1/post/update:
    put:
      consumes:
      - application/json
      description: "this \t\tupdating information of post"
      parameters:
      - description: Post
        in: body
        name: postbody
        required: true
        schema:
          $ref: '#/definitions/models.PostUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.PostResp'
        "400":
          description: Bad Request
          schema:
            type: boject
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: "Update \t\tPost"
      tags:
      - Post
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"